Add files via upload
Two routes. "Employees" is a GET route that displays all the data from the Tiger database in Mongo. "Hire" is a POST route takes employees data (json format) as input. Adds it to the Tiger database on a server (like Postman). I just don't know how the front-end will get the modified data, while it is still in the server. Try these routes out in Postman.